requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
yfinance>=0.2.50
pandas==2.2.2
mplfinance==0.12.10b0
numpy<2.0.0
matplotlib==3.9.0
google-generativeai==0.7.2
requests==2.32.3

sse-starlette
